[Baekjoon-11478] 서로 다른 부분 문자열의 개수
